de2raw now implies 320x240 resolution
This is to make it easier to teach with in courses like CSCE313. The
previous behavior (with the 16 byte header) is now available via the
de2rawh format.

This commit also introduces the convention of using the extension d2h
with de2rawh files, and d2r with de2raw files.